Upload

+++ src/main/java/model/TrinhDoHocVan.java
@@ -1,24 +1,24 @@
 package model;
 
 public class TrinhDoHocVan {
-    private int maTDHV;
+    private long maTDHV;
     private String TTDHV;
     private String chuyenNganh;
 
     public TrinhDoHocVan() {
     }
 
-    public TrinhDoHocVan(int maTDHV, String TTDHV, String chuyenNganh) {
+    public TrinhDoHocVan(long maTDHV, String TTDHV, String chuyenNganh) {
         this.maTDHV = maTDHV;
         this.TTDHV = TTDHV;
         this.chuyenNganh = chuyenNganh;
     }
 
-    public int getMaTDHV() {
+    public long getMaTDHV() {
         return maTDHV;
     }
 
-    public void setMaTDHV(int maTDHV) {
+    public void setMaTDHV(long maTDHV) {
         this.maTDHV = maTDHV;
     }
